+++ NEWS.md
@@ -1,6 +1,9 @@
 # Version (development version)
 
- * ...
+## Miscellaneous
+
+ * Pico username is now inferred using the new `ssh pico.sh user`
+   endpoint.
 
 
 # Version 0.2.0 [2025-08-13]